+++ Kommunicate/Classes/URLBuilder.swift
@@ -13,6 +13,10 @@ final class URLBuilder {
     private var components = URLComponents()
     private var pathComponents = [String]()
 
+    static var kommunicateBotsApi: URLBuilder {
+        return URLBuilder(host: "bots.kommunicate.io")
+    }
+
     static var kommunicateApi: URLBuilder {
         return URLBuilder(host: "api.kommunicate.io")
     }
@@ -62,3 +66,12 @@ final class URLBuilder {
         return self
     }
 }
+
+extension URLBuilder {
+
+    /// Bot Detail url builder
+    static func botDetail(for applicationKey: String, botId: String) -> URLBuilder {
+        return URLBuilder.kommunicateBotsApi.add(paths: ["application", applicationKey, "bot", botId])
+    }
+}
+