+++ Kommunicate/Classes/Views/BotCharacterLimitView.swift
@@ -1,38 +1,36 @@
 //
-//  AwayMessageView.swift
+//  BotCharacterLimitView.swift
 //  Kommunicate
 //
-//  Created by Mukesh on 08/01/19.
+//  Created by Sunil on 05/08/20.
 //
 
 import Foundation
-import ApplozicSwift
-
-/// A view to show away message. It has message label and dotted line view.
-class AwayMessageView: UIView {
+class BotCharacterLimitView: UIView {
 
     enum ConstraintIdentifier: String {
-         case awayMessageViewHeight
+        case botCharacterLimitViewHeight
     }
 
     struct Padding {
-        struct DottedLineView {
-            static let leading: CGFloat = 12.0
-            static let trailing: CGFloat = 23.0
-        }
-
         struct MessageLabel {
             static let top: CGFloat = 5.0
             static let leading: CGFloat  = 20.0
             static let trailing: CGFloat = 20.0
+            static let height: CGFloat = 70.0
+        }
+        struct DottedLineView {
+            static let leading: CGFloat = 12.0
+            static let trailing: CGFloat = 23.0
         }
     }
 
     private let messageLabel: UILabel = {
         let label = UILabel(frame: CGRect.zero)
-        label.font = UIFont(name: "HelveticaNeue-Light", size: 14)
+        label.font = UIFont(name: "HelveticaNeue", size: 14)
         label.contentMode = .center
         label.textAlignment = .center
+        label.backgroundColor = .clear
         label.textColor = UIColor(netHex: 0x676262)
         label.numberOfLines = 4
         return label
@@ -48,10 +46,13 @@ class AwayMessageView: UIView {
         shapeLayer.path = CGMutablePath()
         return shapeLayer
     }()
+
     private let dottedLineViewHeight: CGFloat = 1.0
     lazy private var dottedLineHeightAnchor = dottedLineView.heightAnchor.constraint(equalToConstant: 0)
 
-    override init(frame: CGRect) {
+    lazy private var messageLabelHeightAnchor = messageLabel.heightAnchor.constraint(equalToConstant: 0)
+
+    public override init(frame: CGRect) {
         super.init(frame: frame)
         setupViews()
     }
@@ -78,7 +79,7 @@ class AwayMessageView: UIView {
         dottedLayer.path = path
     }
 
-    func showMessage(_ flag: Bool) {
+    func showDottedLine(_ flag: Bool) {
         dottedLineHeightAnchor.constant = flag ? dottedLineViewHeight:0
         dottedLayer.lineWidth = flag ? dottedLineViewHeight:0
     }
@@ -91,12 +92,30 @@ class AwayMessageView: UIView {
             $0.leading == leadingAnchor + Padding.DottedLineView.leading
             $0.trailing == trailingAnchor - Padding.DottedLineView.trailing
         }
-
+        messageLabelHeightAnchor.isActive = true
         messageLabel.layout {
-            $0.bottom == bottomAnchor
             $0.top == dottedLineView.bottomAnchor + Padding.MessageLabel.top
             $0.leading == leadingAnchor + Padding.MessageLabel.leading
             $0.trailing == trailingAnchor - Padding.MessageLabel.trailing
         }
+    }
+
+    func hideView(hide:Bool) {
+        showDottedLine(!hide)
+        messageLabelHeightAnchor.constant = hide ? 0 : height()
+    }
+
+    func height() -> CGFloat {
+        return Padding.MessageLabel.height + Padding.MessageLabel.top
+    }
+}
+
+extension BotCharacterLimitView: Localizable {
+    enum LocalizedText {
+        static private let filename = Kommunicate.defaultConfiguration.localizedStringFileName
+        static let botCharLimit = localizedString(forKey: "BotCharLimit", fileName: filename)
+        static let removeCharMessage = localizedString(forKey: "RemoveCharMessage", fileName: filename)
+        static let remainingCharMessage = localizedString(forKey: "RemainingCharMessage", fileName: filename)
+
     }
 }