+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [Unreleased]
 
+### Enhancements
+- [CM-289]Add restriction for adding more than 256 chars in case the conversation is assigned to a Dialogflow bot.
+
 ## [5.6.0] - 2020-08-14
 
 ### Enhancements